requirements.txt
flask==3.0.2
openai==1.12.0
anthropic==0.18.1
selenium==4.18.1
webdriver-manager==4.0.1
python-dotenv==1.0.1
nltk==3.8.1 